+++ mysql_db_dump.py
@@ -299,7 +299,7 @@ def run_program(args_array, opt_arg_list, opt_dump_list, **kwargs):
 
     err_sup = args_array.get("-w", False)
     dump_db(dump_cmd, db_list, compress, dmp_path, err_sup=err_sup,
-            mail=mail)
+            mail=mail, use_mailx=args_array.get("-u", False))
     mysql_libs.disconnect(server)